Update sample-config.md
I would remove the colon (:) from all references to key names. I just removed a couple. 
The colon is a yaml syntax token, so it's not part of a key's name. Also, in some keys in your config files do not have values (like checkout) and do not need the colon token.